proxyRootResolver
