+++ src/Data/Morpheus/Server/Deriving/Resolve.hs
@@ -1,15 +1,12 @@
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 
 module Data.Morpheus.Server.Deriving.Resolve
   ( statelessResolver,
     RootResolverConstraint,
-    fullSchema,
     coreResolver,
   )
 where
@@ -20,7 +17,6 @@ import Data.Functor.Identity (Identity (..))
 import Data.Morpheus.Core
   ( runApi,
   )
-import Data.Morpheus.Internal.Utils (resolveUpdates)
 import Data.Morpheus.Server.Deriving.Channels (ChannelCon)
 import Data.Morpheus.Server.Deriving.Encode
   ( EncodeCon,
@@ -28,9 +24,8 @@ import Data.Morpheus.Server.Deriving.Encode
   )
 import Data.Morpheus.Server.Deriving.Introspect
   ( IntroCon,
-    introspectObjectFields,
+    fullSchema,
   )
-import Data.Morpheus.Server.Types.GQLType (GQLType (CUSTOM))
 import Data.Morpheus.Types
   ( RootResolver (..),
   )
@@ -41,32 +36,18 @@ import Data.Morpheus.Types.IO
   )
 import Data.Morpheus.Types.Internal.AST
   ( CONST,
-    FieldsDefinition,
     MUTATION,
-    OUT,
     QUERY,
     SUBSCRIPTION,
-    Schema (..),
-    TypeContent (..),
-    TypeDefinition,
-    TypeName,
-    ValidValue,
-    initTypeLib,
-    mkType,
+    Schema,
+    VALID,
+    Value,
   )
 import Data.Morpheus.Types.Internal.Resolving
-  ( Eventless,
-    Resolver,
+  ( Resolver,
     ResponseStream,
     ResultT (..),
     cleanEvents,
-  )
-import Data.Proxy (Proxy (..))
-
-type IntrospectConstraint m event query mutation subscription =
-  ( IntroCon (query (Resolver QUERY event m)),
-    IntroCon (mutation (Resolver MUTATION event m)),
-    IntroCon (subscription (Resolver SUBSCRIPTION event m))
   )
 
 type OperationConstraint operation event m a =
@@ -94,7 +75,7 @@ coreResolver ::
   RootResolverConstraint m event query mut sub =>
   RootResolver m event query mut sub ->
   GQLRequest ->
-  ResponseStream event m ValidValue
+  ResponseStream event m (Value VALID)
 coreResolver root request =
   validRequest
     >>= execOperator
@@ -104,49 +85,3 @@ coreResolver root request =
     validRequest = cleanEvents $ ResultT $ pure $ fullSchema $ Identity root
     --------------------------------------
     execOperator schema = runApi schema (deriveModel root) request
-
-fullSchema ::
-  forall proxy m event query mutation subscription.
-  (IntrospectConstraint m event query mutation subscription) =>
-  proxy (RootResolver m event query mutation subscription) ->
-  Eventless (Schema CONST)
-fullSchema _ = querySchema >>= mutationSchema >>= subscriptionSchema
-  where
-    querySchema =
-      resolveUpdates (initTypeLib (operatorType fields "Query")) types
-      where
-        (fields, types) =
-          introspectObjectFields
-            (Proxy @(CUSTOM (query (Resolver QUERY event m))))
-            ("type for query", Proxy @(query (Resolver QUERY event m)))
-    ------------------------------
-    mutationSchema lib =
-      resolveUpdates
-        (lib {mutation = maybeOperator fields "Mutation"})
-        types
-      where
-        (fields, types) =
-          introspectObjectFields
-            (Proxy @(CUSTOM (mutation (Resolver MUTATION event m))))
-            ( "type for mutation",
-              Proxy @(mutation (Resolver MUTATION event m))
-            )
-    ------------------------------
-    subscriptionSchema lib =
-      resolveUpdates
-        (lib {subscription = maybeOperator fields "Subscription"})
-        types
-      where
-        (fields, types) =
-          introspectObjectFields
-            (Proxy @(CUSTOM (subscription (Resolver SUBSCRIPTION event m))))
-            ( "type for subscription",
-              Proxy @(subscription (Resolver SUBSCRIPTION event m))
-            )
-    maybeOperator :: FieldsDefinition OUT CONST -> TypeName -> Maybe (TypeDefinition OUT CONST)
-    maybeOperator fields
-      | null fields = const Nothing
-      | otherwise = Just . operatorType fields
-    -------------------------------------------------
-    operatorType :: FieldsDefinition OUT CONST -> TypeName -> TypeDefinition OUT CONST
-    operatorType fields typeName = mkType typeName (DataObject [] fields)
